requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
